refactor(api): split userBike spot routes into dedicated module
